% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_graph.R
\name{choose}
\alias{choose}
\title{Chooses a specific entry in the APSP matrix.}
\usage{
choose(APSP, first, second)
}
\arguments{
\item{APSP}{the All-Pairs-Shortest-Paths matrix obtained by \link{get_apsp}}

\item{first}{string; name of first reader}

\item{second}{string; name of second reader}
}
\value{
the specific entry in APSP(first, second)
}
\description{
Chooses a specific entry in the APSP matrix.
}
