% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{cleanAge}
\alias{cleanAge}
\title{clean the age dataframe}
\usage{
cleanAge(age)
}
\arguments{
\item{age}{Data frame obtained by using \code{\link{readAge}}}
}
\value{
mice ID and their ages as a dataframe
}
\description{
clean the age dataframe
}
