% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{cleanRawData}
\alias{cleanRawData}
\title{cleans the data, creates first new columns}
\usage{
cleanRawData(
  data,
  phases = list(light = 0:11, dark = setdiff(0:23, 0:11)),
  groups
)
}
\arguments{
\item{data}{the raw data}

\item{phases}{list of different phases, each consisting of a vector of hours.}

\item{groups}{list of different groups, each group being a vector of strings.}
}
\value{
cleaned data
}
\description{
cleans the data, creates first new columns
}
