% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColonyRack.R
\docType{package}
\name{ColonyRack}
\alias{ColonyRack}
\title{ColonyRack: a Package for analyzing colonyrack data.}
\description{
The aim of this package is to process data from RFID tag readers and save the intermediate results in separate files to be able to track large numbers of mice.

To learn more about ColonyRack, start with the vignettes: browseVignettes(package = "ColonyRack")

To run the the package functions properly, use the attached markdown. Go to the \code{Files} window in RStudio and inside your project
directory and open the directory \code{vignettes}. There you will find a file named \code{meta_exploration.Rmd}.
Click on it and the vignette will be loaded to RStudio.
You can see that the vignette is just a Rmarkdown file.
All the details of the implemented complex statistical algorithms are described in it.
}
\author{
Mohammad Orabe, \strong{orabe.mhd@gmail.com}
}
