% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{compute_or_read}
\alias{compute_or_read}
\title{Reads the wanted data, or compute it itself, if needed.
Useful on big data frames.}
\usage{
compute_or_read(data, to.compute, filename = NA, force.compute = FALSE)
}
\arguments{
\item{data}{data to use in the toCompute function}

\item{to.compute}{function to compute new data.frame if file not found}

\item{filename}{file to use if exists.
If NA, the user has to choose existent file}

\item{force.compute}{TRUE if computation is wanted, even if file exists}
}
\value{
either computer or read data.frame
}
\description{
Reads the wanted data, or compute it itself, if needed.
Useful on big data frames.
}
