% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{count_visited_cages_per_interval}
\alias{count_visited_cages_per_interval}
\title{Count the number of visited cages in time intervals}
\usage{
count_visited_cages_per_interval(
  data,
  first_bedding_day,
  interval_type = "minute"
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}

\item{first_bedding_day}{The date of the first bedding date as a string (e.g. \code{"2020-08-25"})}

\item{interval_type}{\code{minute} or \code{hour}. Specify how the time interval should be grouped.}
}
\description{
For each day, the length of the time interval can be specified (either per minute or hour).
Note that the date on the \code{Time} column is the \code{today()} date. This date is created for plotting purposes.
However the time on Time column is correct
}
