% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{generate_roaming.entropy}
\alias{generate_roaming.entropy}
\title{Generates roaming entropy data.frame}
\usage{
generate_roaming.entropy(data, cage.values)
}
\arguments{
\item{data}{data.frame with Date, Cage, Group, IdLabel, Duration information.
the Date, IdLabel, Cage are used for grouping,
Duration for the actual roaming entropy.}
}
\value{
data.frame with roaming entropy per date, per mouse or group.
}
\description{
Generates roaming entropy data.frame
}
