% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_24hour_tour_processed}
\alias{get_24hour_tour_processed}
\title{Get the 24hour tour processed data.}
\usage{
get_24hour_tour_processed(data, APSP, cages, connections, coords)
}
\arguments{
\item{data}{Dataframe obtained by using \code{\link{get_24hour_tour_raw}}}

\item{APSP}{All-Pairs-Shortest-Path dataframe obtained by using
\code{\link{get_apsp}}}

\item{cages}{vector of cage-labels}

\item{connections}{vector of connection-labels}
}
\value{
the 24hour tour processed data (see other documentation for details)
}
\description{
Add the segment-data (cage, distance, duration etc.).
}
\details{
\itemize{
\item \code{Segment_type} tells how many cages the mouse crossed at least (and
most probably),
\item \code{Cage_type} tells whether the Cage is an actual Cage (1) or a level
connection (2),
\item One (linear) \code{Crossing} is the traversing of multiple Readers that
are on the same $x or $y level (computed using \code{\verb{\%linear\%}}).
}
}
