% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_24hour_tour_raw}
\alias{get_24hour_tour_raw}
\title{Get the 24hour tour raw data. That is, with Day and double reads information
added.}
\usage{
get_24hour_tour_raw(data, day_start)
}
\arguments{
\item{data}{Dataframe obtained by using \code{\link{cleanRawData}} on
colonyrack-rawdata}

\item{day_start}{int between 0 and 23, set the hour when a day should start}
}
\value{
the 24hour tour raw data, that is,
the input dataframe, but with the \code{Day} and \code{DoubleRead} columns
}
\description{
Get the 24hour tour raw data. That is, with Day and double reads information
added.
}
