% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_graph.R
\name{get_apsp}
\alias{get_apsp}
\title{Computes the All-Pairs-Shortest-Paths (short: APSP) matrix for a given graph.}
\usage{
get_apsp(G)
}
\arguments{
\item{G}{a graph (that is, an adjacency list Dataframe) obtained by
\link{read_adj_list}}
}
\value{
APSP matrix as a dataframe.
The entries are the paths (vector of cage names), "0" (zero) if the
readers are direct neighbors, or "Inf" if the readers are not connected
}
\description{
Computes the All-Pairs-Shortest-Paths (short: APSP) matrix for a given graph.
}
\details{
The output APSP matrix has as entries (strings):
\itemize{
\item "0" ; same reader,
\item "Inf" (infinity) ; The readers are not connected,
\item cage name ; The readers are neighbors, using some cage,
\item any integer greater than 1, as a string ;
The readers are connected, but the distance is at least 2.
}
}
