% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_cage}
\alias{get_cage}
\title{cage detection used for V1-3D-Cage
TODO maybe use function(...){case_when(...,TRUE~NA)}}
\usage{
get_cage(from, to, floor.max = 3)
}
\description{
cage detection used for V1-3D-Cage
TODO maybe use function(...){case_when(...,TRUE~NA)}
}
