% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_cages_perinterval}
\alias{get_cages_perinterval}
\title{summarizes the cages a mouse was in at the same time interval into one column}
\usage{
get_cages_perinterval(all_times, data, time.interval)
}
\arguments{
\item{all_times}{list with the timestamps for the intervals as all_times}

\item{data}{data.frame with cage visits (one cage per visit)}

\item{time.interval}{the interval duration}
}
\value{
a data.frame with cages per interval information.
TODO: IT'S NOT WORKING THAT WELL YET!
}
\description{
summarizes the cages a mouse was in at the same time interval into one column
}
