% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_coords}
\alias{get_coords}
\title{Generate the reader-coords-table.}
\usage{
get_coords(data)
}
\arguments{
\item{data}{the raw data (obtained using \code{\link{readData}})}
}
\value{
Dataframe with (reader, x, y, z) as columns
}
\description{
Generate the reader-coords-table.
}
