% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_crossing_summary}
\alias{get_crossing_summary}
\title{Create locomotion summary for each crossing length}
\usage{
get_crossing_summary(data)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}
}
\description{
For each crossing, summarize distance, duration and speed.
}
