% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_daily_mean_speed}
\alias{get_daily_mean_speed}
\title{Calculate the mean speed per day for each mouse}
\usage{
get_daily_mean_speed(data, phase = NA)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_speed}}}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}
}
\description{
Calculate the mean speed per day for each mouse
}
