% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_date_ids}
\alias{get_date_ids}
\title{gets the id of every date-day in all_times.}
\usage{
get_date_ids(all_times)
}
\arguments{
\item{all_times}{list with dates with regular interval}
}
\value{
a data.fram with one column - the id. (maybe list would be better)
}
\description{
gets the id of every date-day in all_times.
}
