% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_dateReaderData}
\alias{get_dateReaderData}
\title{generates dateReaderData}
\usage{
get_dateReaderData(data)
}
\arguments{
\item{data}{data that was cleaned with cleanRawData function}
}
\value{
dateReaderData
}
\description{
generates dateReaderData
}
