% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_dirsave}
\alias{get_dirsave}
\title{lets the user choose a results directory, generates subdirectories and
returns an environment containing the directory and the subdirectories.}
\usage{
get_dirsave(dir_name = NA)
}
\arguments{
\item{dir_name}{will be read if given, otherwise dialog GUI}
}
\value{
Environment containing the chosen directory and the subdirectories
}
\description{
lets the user choose a results directory, generates subdirectories and
returns an environment containing the directory and the subdirectories.
}
