% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_dist_5m}
\alias{get_dist_5m}
\title{Calculate the traveled distance in cages every 5 minutes}
\usage{
get_dist_5m(data)
}
\arguments{
\item{data}{Dataframe obtained by using \code{\link{get_24hour_tour_processed}}}
}
\description{
Calculate the traveled distance in cages every 5 minutes
}
