% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_distance}
\alias{get_distance}
\title{Calculate traveled distance per day and hour}
\usage{
get_distance(data, phase = NA, method)
}
\arguments{
\item{data}{Dataframe obtained by using \code{\link{get_24hour_tour_processed}}}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}

\item{method}{Specification of the length of the time interval for summing up the distances. Use either \code{perHour} or \code{perDay}.}
}
\description{
Create a data frame that contains total distance values from different phases for each mouse.
}
