% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_distribution_duration}
\alias{get_distribution_duration}
\title{Calculate the distribution of duration in each cage.}
\usage{
get_distribution_duration(data)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}
}
\description{
Calculate the mean, standard deviation, minimum and maximum time spent in each cage for each mouse in day and dark and light phases.
}
