% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_double_reads}
\alias{get_double_reads}
\title{Get the double reads in the tour data.}
\usage{
get_double_reads(data)
}
\arguments{
\item{data}{Dataframe obtained by using \code{\link{cleanRawData}} on
colonyrack-rawdata}
}
\value{
the input dataframe, but with the \code{DoubleRead} column
}
\description{
Get the double reads in the tour data.
}
\details{
\itemize{
\item We look at each separate mouse, call it M.
\item If there is a new event with M, while an earlier event with M has not
finished yet, we call it a \emph{double read}.
\item The number tells us what consecutive double read the event is.
}
}
