% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_duration_inCageLevel}
\alias{get_duration_inCageLevel}
\title{Create total duration with different interval of time length}
\usage{
get_duration_inCageLevel(data, phase = NA, time_window_type)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}

\item{time_window_type}{How the duration should be calculated. Use \code{perDay} for total time spent in cages per day and \code{perHour} for duration per hour.}
}
\description{
For each mouse calculate the time spent in each cage and level (colony rack level) per day or per hour.
}
