% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_duration_inCages}
\alias{get_duration_inCages}
\title{Calculate total duration with different interval of time length}
\usage{
get_duration_inCages(data, time_window_type, phase = NA)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}

\item{time_window_type}{How the duration should be calculated. Use \code{perDay} for total time spent in cages per day and \code{perHour} for duration per hour.}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}
}
\description{
Create a data frame that contains total time spent in cages (no tubes) for each mouse and phase.
For each mouse calculate the time spent in each cage and level (colony rack level) per day or per hour.
Note that the date on the \code{Time} column is the \code{today()} date. This date is created for plotting purposes.
}
