% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_duration_overlap}
\alias{get_duration_overlap}
\title{Calculate the overlap of timespent in cages between all mice permutations}
\usage{
get_duration_overlap(data)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}
}
\description{
The function sums up all duration overlaps (time spent in same cages) between all mice of different lengths.
}
