% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_entropy_duration}
\alias{get_entropy_duration}
\title{Calculate the duration entropy value in each cage.}
\usage{
get_entropy_duration(data)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}
}
\description{
Calculate the duration entropy value in each cage.
}
