% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_locations}
\alias{get_locations}
\title{generates table with mice-cage information.}
\usage{
get_locations(data, reader_cage = get_cage)
}
\arguments{
\item{data}{data that was cleaned with cleanRawData before.}
}
\description{
generates table with mice-cage information.
}
