% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_locomotion_distance}
\alias{get_locomotion_distance}
\title{Calculate the traveled distance for each mouse and phase.}
\usage{
get_locomotion_distance(data)
}
\arguments{
\item{data}{Dataframe obtained by using \code{\link{get_24hour_tour_processed}}}
}
\value{
Data frame contains the total traveled distance, light and dark phase for each mouse.
}
\description{
Calculate the traveled distance for each mouse and phase.
}
