% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_mean_duration_cageLevelCategory}
\alias{get_mean_duration_cageLevelCategory}
\title{Calculate the mean duration in cage categories}
\usage{
get_mean_duration_cageLevelCategory(
  data,
  water_cages,
  food_cages,
  sleeping_cages
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}

\item{water_cages}{Cages that with a water bottle for drinking}

\item{food_cages}{Cages that have food}

\item{sleeping_cages}{Cages that are used for resting or sleeping}
}
\description{
For each mouse, calculate the mean time spent in each cage category (food, water, resting cages) in each phase and cage level
}
