% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_mean_hourlyDistance_perDay}
\alias{get_mean_hourlyDistance_perDay}
\title{Calculate the mean distance per hour for each day}
\usage{
get_mean_hourlyDistance_perDay(data)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_distance}}}
}
\description{
Calculate the mean distance per hour for each day
}
