% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_mean_pianka_index}
\alias{get_mean_pianka_index}
\title{Calculate the mean pianka index for each mouse}
\usage{
get_mean_pianka_index(data, mice_age)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_pairwise_niche_overlap}}}

\item{mice_age}{Data frame contains two columns: The mice ID \code{IdLabel} and their age \code{Age_in_months}}
}
\description{
Mean spatial overlap (Pianka’s index) is a measure of the degree of spatial overlap between
the different individuals. To obtain the mean Pianka’s symmetrical index of niche overlap (Pianka 1973)
the overlap for each mouse with each of the other mice was calculated and then a mean value was
determined from these data for each individual.
The maximum value of the index is 1 and indicates complete overlap,
while the minimum value 0 indicates no overlap.
}
\details{
Reference: Nachev, Vladislav Nikolaev. "Cognition mediated floral evolution." (2014).
}
