% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_mice_per_group}
\alias{get_mice_per_group}
\title{Get a list of groups and for each the vector of mice in the group.}
\usage{
get_mice_per_group(data, groupnames)
}
\arguments{
\item{data}{Dataframe obtained by using \code{\link{cleanRawData}} on
colonyrack-rawdata}

\item{groupnames}{list or vector with names of the groups to be found in data}
}
\value{
list of vectors. One list-entry per group, and the vector contains
the mice ids that are in the group.
}
\description{
Get a list of groups and for each the vector of mice in the group.
}
