% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_pairwise_niche_overlap}
\alias{get_pairwise_niche_overlap}
\title{Calculate the niche overlap value for each pair of mice}
\usage{
get_pairwise_niche_overlap(data, method)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}

\item{method}{Specify on what basis should the overlap be calculated. Use \code{cage_visits} to calculate the niche overlap value
based on the number of visited and \code{duration} for overlap based time spent in cages.}
}
\description{
Spatial overlap (Pianka’s index) is a measure of the degree of spatial overlap between
the different individuals. To obtain Pianka’s symmetrical index of niche overlap (Pianka 1973)
the overlap for each mouse with each of the other mice was calculated.
The maximum value of the index is 1 and indicates complete overlap,
while the minimum value 0 indicates no overlap.
}
\details{
Reference: Nachev, Vladislav Nikolaev. "Cognition mediated floral evolution." (2014).
}
