% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_speed}
\alias{get_speed}
\title{Calculate the speed of each mouse}
\usage{
get_speed(data)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}
}
\description{
The speed is determined by dividing the traveled distance between two cages/tubs and the duration (time spent in cage).
}
