% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_total_duration}
\alias{get_total_duration}
\title{Calculate the duration in each cage.}
\usage{
get_total_duration(data)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}
}
\description{
For each mouse and day, and phase (dark and light) calculate the time spent in each cage.
}
