% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_transition_distr}
\alias{get_transition_distr}
\title{Calculate the transition distribution in UP/Down tubes
For each day Count the number of visited tubes between cage levels (without linear connection on the same level where the height is 0)}
\usage{
get_transition_distr(data)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}
}
\description{
Calculate the transition distribution in UP/Down tubes
For each day Count the number of visited tubes between cage levels (without linear connection on the same level where the height is 0)
}
