% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{lag_generator}
\alias{lag_generator}
\title{small helping function-generator for getting the lag of some value}
\usage{
lag_generator()
}
\description{
small helping function-generator for getting the lag of some value
}
