% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{manipulatedDate_distance_hourly}
\alias{manipulatedDate_distance_hourly}
\title{Calculate traveled distance per hour and and manipulate the DateTime_in (Time) to make a stacked bar.}
\usage{
manipulatedDate_distance_hourly(data, phase = NA)
}
\arguments{
\item{data}{Dataframe obtained by using \code{\link{get_24hour_tour_processed}}}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}
}
\description{
Create a data frame that contains total distance values from different phases for each mouse.
}
