% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_cage_distribution_cutOutiers}
\alias{plot_cage_distribution_cutOutiers}
\title{Plot distribution of time spent in cages as boxplot and cut outliers values}
\usage{
plot_cage_distribution_cutOutiers(
  data,
  plot_file_name,
  plot_height,
  plot_width,
  time_window_type,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_duration_inCageLevel}}}

\item{plot_file_name}{File name to create on disk}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{time_window_type}{How the duration should be calculated. Use \code{perDay} for total time spent in cages per day and \code{perHour} for duration per hour.}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Duration values are plotted per hour or per day. Each panel includes data for one level in the colony rack setup (from down to top). Colors indicate the mice age.
Cages are presented on the x-axis.
The figure is saved as a PNG file in the result environment.
}
