% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_cageVisits_after_bedding_change}
\alias{plot_cageVisits_after_bedding_change}
\title{Plot and save the number of cage visits in time intervals}
\usage{
plot_cageVisits_after_bedding_change(
  data,
  plot_file_name,
  mouse,
  week,
  from,
  to,
  graph_type = "smooth",
  span = 0.75,
  plot_width,
  plot_height,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{count_visited_cages_per_interval}}}

\item{plot_file_name}{File name to create on disk}

\item{week}{Use \code{all} to include data from all weeks or enter the number of the week as an integer number}

\item{from}{Starting time of the interval (e.g. from="06:00")}

\item{to}{Ending time of the interval (e.g. to="13:00")}

\item{graph_type}{Defines the plot type. It can be either \code{smooth}, \code{line} or \code{point}}

\item{span}{Floating number that controls the amount of smoothing for the default loess smoother. Smaller numbers produce wigglier lines, larger numbers produce smoother lines.}

\item{plot_width}{Plot width}

\item{plot_height}{Plot height}

\item{DPI}{Plot resolution(Dots per Inch)}

\item{mouseID}{Choose one mouse from the entire data set. The \code{mouseID} must be a \code{String} value.}
}
\description{
For one specific mouse and each day, plot the total number of visited cages as a smoothed line.
Each panel represents data for one week. Data points (line/dots) indicate the days of each week. The first day in each panel is the day in which the bedding has been changed.
The input data, that include the number of cage visits, can be be both grouped per hour or minute. See \code{\link{help(count_visited_cages_per_interval)}} for more information.
Use \code{span} to control the "wiggliness" of the default loess smoother.
The time interval of interest can be specified using the parameters \code{from} and \code{to} which are the upper and lower time limits (in hour)
The figures are saved as PNG files in the corresponding phase-folder in the result environment.
}
