% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_distance_daily_xAxisAge}
\alias{plot_distance_daily_xAxisAge}
\title{Plot and save traveled distance for each week.}
\usage{
plot_distance_daily_xAxisAge(
  data,
  plot_file_name,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Dataframe obtained by using \code{\link{get_distance}}}

\item{plot_file_name}{File name to create on disk}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Present the distribution of daily traveled distance for each week as a boxplot.
Each panel include one week. Mice age is plotted on the x-axis. Colors indicate the mice ID.
The figure is saved as a PNG file in the result environment.
}
