% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_distance_hourly_xAxisDay_smooth}
\alias{plot_distance_hourly_xAxisDay_smooth}
\title{Plot and save traveled distance for each week.}
\usage{
plot_distance_hourly_xAxisDay_smooth(
  data,
  plot_file_name,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_distance}}}

\item{plot_file_name}{File name to create on disk}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Present the distribution of traveled distance per hour for day and week as a smoothed line
Each panel includes data for one mouse. Days are plotted on the x-axis.
Colors indicate the days of the week. Each column of panels describes the data for one week.
Each row of panels describes data for one mouse.
The figure is saved as a PNG file in the result environment.
}
