% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_distance_hourly_xAxisDay_stackedBar}
\alias{plot_distance_hourly_xAxisDay_stackedBar}
\title{Plot and save traveled distance for each day}
\usage{
plot_distance_hourly_xAxisDay_stackedBar(
  data,
  plot_file_name,
  plot_height,
  plot_width,
  nr_xTicks,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{manipulatedDate_distance_hourly}}}

\item{plot_file_name}{File name to create on disk}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{nr_xTicks}{Number of ticks and labels to show on the x-axis}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Present traveled distance per hour for each day as a stacked bar.
Each panel includes data for one mouse. Days are plotted on the x-axis.
Colors indicate the distance value (in meter) for each hour. Each bar presents one day.
The figure is saved as a PNG file in the result environment.
}
