% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_duration_absoluteFreq_freqpoly_tubes}
\alias{plot_duration_absoluteFreq_freqpoly_tubes}
\title{Plot and save the frequency of duration as polygon graphs}
\usage{
plot_duration_absoluteFreq_freqpoly_tubes(
  data,
  plot_file_name,
  max_crossing_length,
  log_scale = F,
  duration_limit_linear_tubes,
  duration_limit_UpDown_tubes,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}

\item{plot_file_name}{File name to create on disk}

\item{max_crossing_length}{Maximum number of crossing to be considered.}

\item{log_scale}{A Boolean parameter for scaling the duration values on the y-axis. Use either \code{TRUE} or \code{FALSE}.}

\item{duration_limit_linear_tubes}{Specify the limit of the duration value on the x-axis when plotting linear tupes}

\item{duration_limit_UpDown_tubes}{Specify the limit of the duration value on the x-axis when plotting up and down tupes}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Plot the duration with absolute frequency for specific number of crossing length (number of continuous linear crossing between cages/tubes).
Only crossing in tubes is considered (no cages). The function can cut out irrelevant duration values using \code{duration_limit_linear_tubes} and \code{duration_limit_linear_tubes}.
Set the parameter \code{log_scaled} to \code{TRUE} to scale the values on the y-axis. The function generatet wo plots for each phase.
The figures are saved as PNG files in the corresponding phase-folder in the result environment.
}
