% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_duration_absoluteFreq_linearCages_hist}
\alias{plot_duration_absoluteFreq_linearCages_hist}
\title{Plot duration histogram for linear cages}
\usage{
plot_duration_absoluteFreq_linearCages_hist(
  data,
  plot_file_name,
  log_scale = F,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}

\item{plot_file_name}{File name to create on disk}

\item{log_scale}{A Boolean parameter for scaling the duration values on the y-axis. Use either \code{TRUE} or \code{FALSE}.}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Plot the duration with absolute frequency for all possible crossing lengths (number of continuous linear crossing between cages/tubes).
Only crossing between cages is considered (no tubes).
Set the parameter \code{log_scaled} to \code{TRUE} to scale the values on the y-axis. The function generates two plots for each phase.
The figures are saved as PNG files in the corresponding phase-folder in the result environment.
}
