% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_duration_absoluteFreq_upDownTube_hist}
\alias{plot_duration_absoluteFreq_upDownTube_hist}
\title{Plot and save duration histogram for connection (tubes)}
\usage{
plot_duration_absoluteFreq_upDownTube_hist(
  data,
  plot_file_name,
  duration_length,
  log_scale = F,
  plot_height,
  plot_width,
  DPI = 150,
  binwidth = binwidth
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}

\item{plot_file_name}{File name to create on disk}

\item{duration_length}{maximum duration length to be considered on the x axis.}

\item{log_scale}{A Boolean parameter for scaling the duration values on the y-axis. Use either \code{TRUE} or \code{FALSE}.}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}

\item{binwidth}{The width of the bins. Can be specified as an integer number The width of the bins. Can be specified as an integer number}
}
\description{
Plot absolute frequency for the duration for only crossing in up and down tubes (no cages or linear tubes).
Set the parameter \code{log_scaled} to \code{TRUE} to scale the values on the y-axis.
The figures are saved as PNG files in the corresponding phase-folder in the result environment.
}
