% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_duration_distr_cageLevelCategory}
\alias{plot_duration_distr_cageLevelCategory}
\title{Plot and save distribution of duration in cages}
\usage{
plot_duration_distr_cageLevelCategory(
  data,
  plot_file_name,
  phase,
  weeks = NA,
  log_scale = TRUE,
  plot_width,
  plot_height,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_mean_duration_cageLevelCategory}}}

\item{plot_file_name}{File name to create on disk}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}

\item{weeks}{Vector of number of weeks to be considered i.g. c(1,5)}

\item{log_scale}{A Boolean parameter for scaling the duration values on the y-axis. Use either \code{TRUE} or \code{FALSE}.}

\item{plot_width}{Plot width}

\item{plot_height}{Plot height}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Plot the mean of time spent in each cage type(food, water, rest cages) for each level in the colony rack setup.
The figure is saved as a PNG file in the result environment.
}
