% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_duration_heatmap_distr}
\alias{plot_duration_heatmap_distr}
\title{Plot distribution of duration as a heatmap}
\usage{
plot_duration_heatmap_distr(
  data,
  dist_type = NA,
  plot_file_name,
  phase = NA,
  plot_height,
  plot_width,
  DPI = 150,
  nr_xTicks
)
}
\arguments{
\item{data}{Data frame obtained by the function \code{\link{get_distribution_duration}}}

\item{dist_type}{Use \code{mean} or \code{sd} for calculating the mean or stadard deviation of the traveled distance}

\item{plot_file_name}{File name to create on disk}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}

\item{nr_xTicks}{Number of ticks and labels to show on the x-axis}
}
\description{
Plot the mean or the standard deviation of time spent in cages for each mouse and day.
Each panel represents a data for one mouse. Color intensity indicate (mean duration or sd value) of time spent in each cage.
Days are plotted on the x-axis.
The figure is saved as a PNG file in the result environment.
}
