% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_duration_heatmap_entropy}
\alias{plot_duration_heatmap_entropy}
\title{Plot duration entropy as a heatmap}
\usage{
plot_duration_heatmap_entropy(
  data,
  plot_file_name,
  phase = NA,
  plot_height,
  plot_width,
  DPI = 150,
  nr_xTicks
)
}
\arguments{
\item{data}{Data frame obtained by the function \code{\link{get_entropy_duration}}}

\item{plot_file_name}{File name to create on disk}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}

\item{nr_xTicks}{Number of ticks and labels to show on the x-axis}
}
\description{
Plot the entropy score value for each mouse in each cage.
Each panel represents a data for one mouse. Color intensity indicate the entropy value in each cage.
Days are plotted on the x-axis.
The figure is saved as a PNG file in the result environment.
}
