% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_duration_relativeFreq_tubes_hist_customized}
\alias{plot_duration_relativeFreq_tubes_hist_customized}
\title{Plot and save duration histogram for connection (tubes)}
\usage{
plot_duration_relativeFreq_tubes_hist_customized(
  data,
  plot_file_name,
  max_duration_length = NULL,
  log_scale = F,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_24hour_tour_processed}}}

\item{plot_file_name}{File name to create on disk}

\item{max_duration_length}{maximum duration length}

\item{log_scale}{A Boolean parameter for scaling the duration values on the y-axis. Use either \code{TRUE} or \code{FALSE}.}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Plot relative frequency for the duration tubes (linear, up and down tubes) for a specific number of crossing lengths.
Set the parameter \code{log_scaled} to \code{TRUE} to scale the values on the y-axis.
The figures are saved as PNG files in the corresponding phase-folder in the result environment.
}
