% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_heatmap_count_cage_visits_hourly}
\alias{plot_heatmap_count_cage_visits_hourly}
\title{Plot and save the number of visited cages as a heatmap}
\usage{
plot_heatmap_count_cage_visits_hourly(
  data,
  plot_file_name,
  phase = NA,
  day,
  nr_xTicks,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_daily_mean_speed}}}

\item{plot_file_name}{File name to create on disk}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}

\item{day}{Specify the day of interest as an integer number}

\item{nr_xTicks}{Time difference between hour-labels to show on the x-axis}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
For each mouse and within a specific day plot the number of visited cages per hour
Color intensity indicates the number of cage visits. Time is on the x-axis. Each panel represents data one cage.
Each row-of-panels refer to one level in the colony rack setup.
The figure is saved as a PNG file in the result environment.
}
