% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_heatmap_duration_inCages_hourly}
\alias{plot_heatmap_duration_inCages_hourly}
\title{Plot and save the duration for one day as a heatmap}
\usage{
plot_heatmap_duration_inCages_hourly(
  data,
  plot_file_name,
  phase = NA,
  day,
  nr_xTicks,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_duration_inCages}}}

\item{plot_file_name}{File name to create on disk}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}

\item{day}{Integer number. Specify the day for which the duration per hour should be plotted.}

\item{nr_xTicks}{Number of ticks and labels to show on the x-axis}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Select one day and plot the time spent in each cage per hour for each mouse. Color intensity represents the duration. Time (in hour) is set on the x-axis.
per hour within a day, for each mouse
The figure is saved as a PNG file in the result environment.
}
