% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_locomotion_distance_heatmap}
\alias{plot_locomotion_distance_heatmap}
\title{Plot and save locomotion-distance as a heatmap and save it as a PNG file.
The figure is saved as a PNG file in the result environment.}
\usage{
plot_locomotion_distance_heatmap(
  data,
  plot_file_name,
  plot_height,
  plot_width,
  DPI = 150,
  nr_xTicks
)
}
\arguments{
\item{data}{Data frame obtained by the function \code{\link{get_locomotion_distance}}}

\item{plot_file_name}{File name to create on disk}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}

\item{nr_xTicks}{}
}
\description{
Plot and save locomotion-distance as a heatmap and save it as a PNG file.
The figure is saved as a PNG file in the result environment.
}
