% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_locomotion_distance_phases_heatmap}
\alias{plot_locomotion_distance_phases_heatmap}
\title{Plot and save locomotion-distance for dark/light/day as a heatmap and save it as a PNG file.
The figure is saved as a PNG file in the result environment.}
\usage{
plot_locomotion_distance_phases_heatmap(
  data,
  plot_file_name,
  plot_height,
  plot_width,
  nr_xTicks,
  DPI = 150
)
}
\arguments{
\item{data}{locomotion dataframe}

\item{plot_file_name}{File name to create on disk}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{nr_xTicks}{Number of ticks and labels to show on the x-axis}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Plot and save locomotion-distance for dark/light/day as a heatmap and save it as a PNG file.
The figure is saved as a PNG file in the result environment.
}
