% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_mean_hourlyDistance_perDay_xAxisDay_bar}
\alias{plot_mean_hourlyDistance_perDay_xAxisDay_bar}
\title{Bar plot for mean traveled distance per hour, per day. Day is on X-axis}
\usage{
plot_mean_hourlyDistance_perDay_xAxisDay_bar(
  data,
  plot_file_name,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_mean_hourlyDistance_perDay}}}

\item{plot_file_name}{File name to create on disk}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Days are plotted on the x-axis. Colors indicate the mice age.Each column of panels includes data for one week.
Each row includes data for one mouse.
The figure is saved as a PNG file in the result environment.
}
