% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_mean_pianka_index}
\alias{plot_mean_pianka_index}
\title{Plot and save the mean pianka's index for each mouse.}
\usage{
plot_mean_pianka_index(
  data,
  plot_file_name,
  subtitle,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_mean_pianka_index}}}

\item{plot_file_name}{File name to create on disk}

\item{subtitle}{should indicate the method (cage visits or duration)}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
The pianka's index values are plotted as a dots.
The maximum value of the index is 1 and indicates complete overlap,
while the minimum value 0 indicates no overlap.
The figure is saved as a PNG file in the result environment.
}
