% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_pairwise_niche_overlap}
\alias{plot_pairwise_niche_overlap}
\title{Plot and save niche-overlap between pairs.}
\usage{
plot_pairwise_niche_overlap(
  data,
  plot_file_name,
  subtitle,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_pairwise_niche_overlap}}}

\item{plot_file_name}{File name to create on disk}

\item{subtitle}{Text for subtitle the figure. Should indicate the method usded (cage visits or duration)}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
The pianka's index values are plotted as a heat map.
The maximum value of the index is 1 and indicates complete overlap,
while the minimum value 0 indicates no overlap.
The figure is saved as a PNG file in the result environment.
}
