% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_speed_daily_xAxisDay_bar}
\alias{plot_speed_daily_xAxisDay_bar}
\title{Plot and save mean speed per day as a bar plot.}
\usage{
plot_speed_daily_xAxisDay_bar(
  data,
  plot_file_name,
  phase = NA,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_daily_mean_speed}}}

\item{plot_file_name}{File name to create on disk}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
For each week and day, speed values of each mouse are plotted.
Each panel includes data for one mouse and one week. Colors indicate the mice age.
Each row of panels represents one mouse. Each column of panels represents one week
Days are presented on the x-axis.
The figure is saved as a PNG file in the result environment.
}
