% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_speed_distribution}
\alias{plot_speed_distribution}
\title{Plot and save speed distribution as a box plot}
\usage{
plot_speed_distribution(
  data,
  plot_file_name,
  phase = NA,
  nr_xTicks,
  plot_height,
  plot_width,
  log_scale = F,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_speed}}}

\item{plot_file_name}{File name to create on disk}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}

\item{nr_xTicks}{Number of ticks and labels to show on the x-axis}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{log_scale}{A Boolean parameter for scaling the duration values on the y-axis. Use either \code{TRUE} or \code{FALSE}.}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Speed values are presented for each mouse per day .
Each panel includes data for one mouse. Colors indicate the mice age.
Days are presented on the x-axis.
The figure is saved as a PNG file in the result environment.
}
