% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_speed_frequency_poly}
\alias{plot_speed_frequency_poly}
\title{Plot and save speed frequency  as polygon graph}
\usage{
plot_speed_frequency_poly(
  data,
  plot_file_name,
  phase = NA,
  crossing_length,
  cage_type,
  height,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_speed}}}

\item{plot_file_name}{File name to create on disk}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}

\item{crossing_length}{integer number indicates to all underlying cross values. Default is all occurrences.}

\item{cage_type}{Integer number. Use 1 to select cages and 2 for tubes}

\item{height}{Indicate the tube type. Use a \code{height} value  greater than 0 for UP-tubes, less than zero for DOWN-tubes and 0 for linear tubes}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
Specify the crossing length by adjusting the \code{crossing_length}. Speed in cages and tubes (linear, up and down) can be plotted.
The figure is saved as a PNG file in the result environment.
}
