% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_transition_tube_xAxisWeek}
\alias{plot_transition_tube_xAxisWeek}
\title{Plot and save weekly transition distribution in UP/Down tubes as a boxplot}
\usage{
plot_transition_tube_xAxisWeek(
  data,
  plot_file_name,
  phase,
  weeks = NULL,
  plot_height,
  plot_width,
  DPI = 150
)
}
\arguments{
\item{data}{Data frame obtained by using \code{\link{get_transition_distr}}}

\item{plot_file_name}{File name to create on disk}

\item{phase}{The \code{dark} or \code{light} phase to be considered. If \code{NA} then the values are calculated for 24 hours.}

\item{weeks}{Vector of number of weeks to be considered i.g. c(1,5)}

\item{plot_height}{Plot height}

\item{plot_width}{Plot width}

\item{DPI}{Plot resolution(Dots per Inch)}
}
\description{
For each week and phase plot the number of visits in tubes where mice can go up and down (linear connections on the same level where height is 0 are not considered).
Each boxplot indicates data for one mouse, which is the number of transitions in tubes for one week.
A box plot is created from 7 days (one week), with each day containing the sum of all visits on that day.
Each column-of-panels represents one mouse, The first row-panels include data for mice that use the tubes to go up (UP-tube);
the second row-of-panels include tubes that are used for down transitions in the colony rack setup.
Colors indicate the tube type number (e.g. \code{Tube 1:2} is the connection between the level 1 and 2 with no movement direction a.k.a. does not mean the movement from level 1 one to level 2).
Values on the x-axis refer to the week number.
The figures are saved as PNG files in the corresponding phase-folder in the result environment.
}
