% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_graph.R
\name{read_adj_list}
\alias{read_adj_list}
\title{Reads the adjacency list of the reader+cage neighborhood.}
\usage{
read_adj_list(filename = NA)
}
\arguments{
\item{filename}{the default path to use. If NA, a dialog box will appear.}
}
\value{
a Dataframe with the read adjacency list (call this a "graph").
}
\description{
Reads the adjacency list of the reader+cage neighborhood.
}
\details{
Expected conventions are:
* no names with "," "." ";",
* no stand-alone integers,
* readers are not "direct neighbors" of each other.
The Graph is given as a list of readers (one row per reader) and the
corresponding list of edges the reader is adjacent to.
The edges are labeled as the cages.
}
