% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{readAge}
\alias{readAge}
\title{read mice ages to merge it with later dailyTourRaw}
\usage{
readAge(filenames = NA)
}
\arguments{
\item{filenames}{will be read if given, otherwise dialog GUI}
}
\value{
some raw data with mice age and ID
}
\description{
read mice ages to merge it with later dailyTourRaw
}
