% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{readData}
\alias{readData}
\title{Reads the chosen data}
\usage{
readData(filenames = NA)
}
\arguments{
\item{filenames}{will be read if given, otherwise dialog GUI}
}
\value{
the raw data
}
\description{
Reads the chosen data
}
