% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{same_reader}
\alias{same_reader}
\title{Returns a function that chooses the "right" cage based on what cage the
mouse was in before.}
\usage{
same_reader(first_cage)
}
\description{
Returns a function that chooses the "right" cage based on what cage the
mouse was in before.
}
