% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{save_24hour_tour}
\alias{save_24hour_tour}
\title{Save the 24 hour tour data (either raw or processed)}
\usage{
save_24hour_tour(data, group_mice, dirsave, type)
}
\arguments{
\item{data}{Dataframe obtained by using \code{\link{get_24hour_tour_raw}}}

\item{group_mice}{list of vectors. One list-entry per group, and the vector
contains the mice ids that are in the group.}

\item{dirsave}{environment containing all the results-directory-paths}

\item{type}{either "raw" or "processed", indicating the type of the table}
}
\description{
Save the 24 hour tour data (either raw or processed)
}
\details{
inside the dirsave$csv tour_raw or tour_processed directory, creates
one subdirectory per group, inside that one subdirectory per mouse in the
group, and inside that subdirectory per mouse, make a subdirectory for days.
and store all csv file per day.
}
