% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{set_day}
\alias{set_day}
\title{Set the day of each entry.}
\usage{
set_day(data, day_start)
}
\arguments{
\item{data}{Dataframe obtained by using \code{\link{cleanRawData}} on
colonyrack-rawdata}

\item{day_start}{int between 0 and 23, set the hour when a day should start}
}
\value{
the input dataframe, but with the \code{Day} column
}
\description{
Set the day of each entry.
}
